add empirical tests to judge sample sizes
Adds a scoreSampleSize() function and accompanying tests to score sample
sizes based on how well they represent heavy hitters in test data of
varying sizes and distributions.

TestSampleSizeVaryingMultiplicities creates test rows representing a
table by filling a slice of length numRows with the numbers [0,
numFrequentVals), with varying multiplicities. The rest of the slice is
then filled with numbers numFrequentVals and up.

TestSampleSizeUniform similarly creates test rows where every value is
"frequent". Each value [0,numFrequentVals) has multiplicity
frequentValMultiplicity.

The scoreSampleSize function runs the sampler on these rows with a
specified sample size. It then counts the number of distinct vals in the
sample that are less than numFrequentVals, as these are the heavy
hitters. The score is computed as the ratio of the number of frequent
vals observed in the sample to the number of frequent vals in the rows.

Useful to eventually dynamically determine sample sizes when creating
table stats.